backend/requirements.txt
# ==============================================================================
# CV AGENT BACKEND REQUIREMENTS
# ==============================================================================
#
# This requirements.txt contains all dependencies tested and working
# for the CV Agent Backend application.
#
# Installation: pip install -r requirements.txt
# ==============================================================================

# ==============================================================================
# CORE FASTAPI AND SERVER DEPENDENCIES
# ==============================================================================
fastapi==0.100.0
uvicorn[standard]==0.21.0
python-dotenv==1.1.0
python-multipart==0.0.20
pydantic==2.11.7

# ==============================================================================
# AI AND MACHINE LEARNING DEPENDENCIES
# ==============================================================================
openai==1.90.0
sentence-transformers==3.0.1
torch==2.2.2
scikit-learn==1.4.0
flashtext==2.7
transformers==4.52.4
huggingface-hub==0.33.0
tokenizers==0.21.1
safetensors==0.5.3

# ==============================================================================
# NLP AND LANGUAGE PROCESSING
# ==============================================================================
spacy==3.7.6
# Note: Download English model with: python -m spacy download en_core_web_sm

# ==============================================================================
# DOCUMENT PROCESSING
# ==============================================================================
pdfplumber==0.11.7
PyPDF2==3.0.1
PyMuPDF==1.25.1
python-docx==1.1.0
docx2txt==0.8
jinja2==3.1.2
beautifulsoup4==4.13.4

# ==============================================================================
# HTTP CLIENTS AND WEB SCRAPING
# ==============================================================================
httpx==0.28.1
requests==2.32.4
pyquery==2.0.1
cssselect==1.3.0
pyppeteer==2.0.0
pyee==11.1.1
fake-useragent==2.2.0
parse==1.20.2
w3lib==2.3.1

# ==============================================================================
# CORE UTILITIES AND DEPENDENCIES
# ==============================================================================
numpy==1.26.4
scipy==1.13.1
tqdm==4.67.1
filelock==3.18.0
packaging==25.0
regex==2024.11.6
fsspec==2025.5.1
joblib==1.5.1
threadpoolctl==3.6.0

# ==============================================================================
# SYSTEM AND PLATFORM DEPENDENCIES
# ==============================================================================
appdirs==1.4.4
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
typing-extensions==4.14.0
urllib3==1.26.20
zipp==3.23.0

# ==============================================================================
# WEBSOCKETS AND ASYNC
# ==============================================================================
websockets==10.4

# ==============================================================================
# ADDITIONAL UTILITIES
# ==============================================================================
distro==1.9.0
jiter==0.10.0
hf-xet==1.1.5